requirements.txt
flask==3.0.3
google-generativeai==0.6.0
qdrant-client==1.8.0
markdown==3.5.0